forkly/backend/requirements.txt
# Forkly Backend Dependencies

# Core Django
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Environment & Configuration
python-dotenv==1.0.0

# Security & Rate Limiting
django-ipware==6.0.0
django-axes==6.3.0

# Caching (opcional)
redis==5.0.1
django-redis==5.4.0

# AI & External APIs
openai==1.12.0
requests==2.31.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0